Represent parenthesised expressions in parse tree
Parenthesisation of subexpressions is important for reconstructing the
source, but the parse tree was losing this information previously. This
commit introduces a ParenthesisedParseNode to represent enclosing an
expression in parentheses and abstracts out the parsing of such
expressions into a new method. In the execution tree, this information
is no longer present, and the expression body is used as before.