Add bracket operators
Bracket operators are mirrored brackets like [ and ] surrounding their
arguments. This commit adds both explicit-receiver (indexing) requests &
definitions and implicit-receiver (constructor) requests and
definitions.

Bracket tokens must start and end with mirroring brackets according to
the Unicode BiDi specification, and may have other operators within the
sequence as well as bracketing characters. Bracket constructs always
come in pairs: an opening bracket, and the closing bracket with the
mirrored characters (if applicable) in reverse order at the end. It is
an error to interleave brackets.