Do not include default methods on all part-objects
The default methods that should be on every object should only be on the
most-inherited object. Otherwise, they will shadow overriding versions
from user superclasses. They are moved out to a separate trait-like
object and made optional when constructing a GraceObject. All user
objects that do not inherit from anything else inherit from this trait
object and obtain the methods in that fashion.