Allow defining Grace extensions to builtins
Numbers and strings can be extended with Grace code defining new methods
or overriding existing methods. A builtins file will define Grace
objects to be used. These objects will inherit from the built-in object,
but are evaluated in a special context.

A builtins file can contain defs of object constructors. These defs must
be named "number" and/or "string". The object constructors will be
evaluated for every number/string created, and will be given the builtin
as their parent under the name "builtin".

The console entry point has an option --builtins-override that takes as
its argument the path to such a file to use.