Allow blocks to be variadic
Blocks should be able to have variadic *parameters as other methods are,
but the parser and runtime representation did not address these. Parsing
variadic parameters is complicated for blocks because they could also be
prefix operators, so many layers of testing are required. At runtime,
the block populates its parameters in scope in the same manner that user
methods do.