Expose parts of compiler to user code
Experimentally, user code will now be able to access parts of the
compiler through the import mechanism. Code can:
    import "!compiler" as compiler
to gain access to an object that exposes the parser and a set of
patterns corresponding to the ParseNode subclasses. The parse method
returns a GraceObjectProxy around the ObjectParseNode given natively.
All ParseNodes accessed will also be proxy objects.