Add experimental if-then specialisation node
The IfThenRequestNode represents an if(...) then { ... } request, but
refrains from constructing and applying a block in the ordinary case. If
any arguments are unusual or their evaluations result in unexpected
values, or if "if-then" resolves to anything other than the built-in
version, the node will permanently change to performing the default
implicit-receiver request behaviour.

In the common case, the node evaluates the condition, and if it is the
True singleton explicitly evaluates the body of the "then" block. If the
block does not declare any local names, no local scope is allocated
either.