Allow conditions on error message definitions
A definition of an error message in the data file will now be able to
specify the conditions under which it should be used. Conditions are
specified as Grace code between |[ and ]| at the beginning of the
message, after the identifying code. Any substitution variables provided
are available in scope. If a condition is not met, the search continues
to the next definition in the file.