Retain final identity of object in part objects
GraceObjects gain a new Identity property, which is set to the final
identity of the object. Without inheritance, this is always "this". With
inheritance, they will be different in the parent part-objects, and
native methods may need to know the final identity as well as the
identity of the part-object involved. In particular, == and != in the
default method set are provided as native code and need access to these
identities to allow "super == self" to be correct.

User code always accesses the "self" name, which has always had the
correct final value, and is not affected by this change.