Require that parent names be used as receivers
The names of inherited parents such as "super" are only permitted as
receivers of method requests, and this was not previously enforced. As
the names are not keywords misuses cannot be detected at the same point
that "outer" misuses are, so a subsequent checking pass after the parse
tree is complete detects the error.