Remove fresh-object "-object" method doubling
Earlier versions of this system created a copy of any method returning a
fresh object with a suffixed part "object", as Minigrace did. This
functionality is no longer required, with the reified requests passing
the inheritance information around, and the double presence of methods
was harmful in some cases. They will no longer be created.