Do not uninitialise a var that already has a value
Because var fields can be assigned before the declaration is
encountered, previously a var could become uninitialised after it held a
value:
    object {
        x := 5
        print(x)
        var x
        print(x)
    }
The above code would print "5", then give an uninitialised read error.
Now, vars will not be given the uninitialised value when the declaration
is encountered if they have already been created, as they will have been
for a field.