Prohibit multiple returns within single method
Previously, a second non-local return within a single method's dynamic
extent would simply have the effect of replacing the return value,
because completion was not marked until the point where the method
actually returned. A subsequent return in a finally block or otherwise
will now cause an error as well as other repeated returns.