Distinguish user, distribution, and default errors
The DefaultErrorMessages.txt file providing the default error messages
can be shadowed with other files providing per-user messages, or an
overlay of messages for a particular distribution. Two files will be
used for these:
* OverlayErrorMessages.txt, in the same place as the default
* UserErrorMessages.txt, in the user import root
A user message will be used if it exists, otherwise a distribution
message if it exists, and finally the default message.