Remove magic !compiler import
With the advent of native-module imports, this hack is no longer
required. Instead, code may import "mwh.nz/kernan/compiler" with the
same effect. The code used to build the module is moved to a new
NativeModules directory, but is not compiled by default in preparation
for an eventual further spin-out. The file must be built as a library
(target:library) and with a reference to GraceLanguage.dll, as for all
native modules.