Treat entry point location as a module root
When a particular file is executed at the console, imports will be
processed relative to its location, in addition to the global and
per-user paths. When the repl is used without a file, imports will be
processed relative to the current directory. The GUI continues to use
the static locations as it has no concept of a file location.