Add minimal resource imports stub
Resource imports, which have the form x/y/z.ext, refer to some sort of
resource rather than Grace code, with the handler determined by the
extension. This commit introduces the concept and a built-in handler for
the ".txt" extension that creates a string. There is no user extension
of the handlers at present.