Generalise handling of >. to operator sequences
>. indicates a request on the result of a generic request, and was
handled by un-lexing a codepoint when it was found. The same situation
could occur at the end of nested generics, however, and this case was
not handled at all. Now, the un-lexing step is generalised to all
generic sequences, and any errors deferred until they are known to have
occurred when an individual illegal token is found.