Parse RHS of type statement as expression with ID
If an identifier token is on the right-hand side of a type statement, it
was previously parsed as an identifier alone, precluding any chaining or
generic arguments. Instead, it will now trigger ordinary expression
parsing (as the {} type special case has been excluded already),
allowing any complex expression to appear there.