Report error on cyclic imports
Module imports are not allowed to be cyclic: a module cannot
(transitively) import itself, in order to ensure that the module object
is initialised before it can be accessed by an import name. Previously a
cyclic import would cause a crash, but now a new error R2011 will be
reported.