Allow encoding codepoints to UTF formats
Codepoints objects gain methods utf8, utf16, and utf32. The utf8 method
returns a byte string, while the others return a view of the code units
of the encoding. Those views can be made concrete with the "be" and "le"
methods for big-endian or little-endian encoding. All of the following
give byte strings:
    x.codepoints.utf8
    x.codepoints.utf16.le
    x.codepoints.utf32.be
The views and byte strings are all immutable.