Reconfigure operator and identifier symbols
The previous definition of operators included many characters that were
not generally recognisable as operators, including alphanumeric symbols.
This commit replaces the definition based on general categories with an
explicit list of blocks to consider operators:
     Mathematical Operators
     Supplemental Mathematical Operators
     Miscellaneous Mathematical Symbols-A
     Miscellaneous Mathematical Symbols-B
     Miscellaneous Symbols and Arrows
     Arrows
     Supplemental Arrows-A
     Supplemental Arrows-B
     Supplemental Technical
     Currency Symbols
     Geometric Shapes
The ASCII keyboard operators, mathematical operators, and punctuation
from Latin-1 are also included.

Codepoints in Symbol, other that are not operators are reclassified as
identifier characters, and may be used for method and variable names.
Other mathematical symbols and punctuation are not reclassified as
identifiers, and are now not considered to have any meaning.