Add -c option to evaluate a single expression
The console entry point will now accept an option `-c CODE` to evaluate
CODE and display the result. Multiple -c options can be provided and
they are executed in order in the same scope, exactly as in the REPL.
The REPL's execution code is abstracted out and used by both the REPL
and the -c option.