Add rudimentary tab completion to REPL
Pressing the tab key will cause the REPL to parse and partially execute
the code on the current line speculatively, and then present options for
method and field names found in the relevant object (or the module
object). Side effects may occur, but for sensible field-access-only
attempts they will not.