Report final object ID in default asString
The default asString method returned a memory address in the
stringification based on the object containing those methods (the
internal true receiver) instead of the self-identity of the object
itself, because it used the wrong kind of DelegateMethodNode. This
commit changes to a different kind of method node and makes flow-on
changes elsewhere necessitated by the signature change in GraceObject.