Add version of IDE that speaks WebSockets
After starting grace --ws, loading web-editor-ws/index.html will connect
to the local socket. Files can be created and edited as usual. Running
and stopping uses the standard button. A module must be built before it
can be imported by another module. Output appears in the console, in
real time, and the dom module creates a new window for the code to
interact with. Most of the new code is in minigrace.js, but other
modifications also exist.