Restrict methods and imports to valid locations
Previously, the parser allowed any kind of statement or declaration to
appear anywhere a statement-level item was expected, but some kinds may
only appear in restricted places: an import statement can only appear at
the top level, for example, and a method cannot contain another method.
This commit ensures that such code is a parse error.