Bind methods onto local scopes at execution tree
While method declarations cannot be nested syntactically, some
constructs (such as generic types) naturally give rise to methods, and
it is more consistent to do so universally than to special-case some
bodies. Given that, method nodes may exist in method and block bodies,
and with this commit they will be bound on the local (dynamic) scope
object, which has the correct accessibility and is capable of responding
to requests as any other surrounding object.