Create conflict methods when multiple inherited
If two or more methods are inherited by the same name, neither will be
in the final object, and a conflict marker will be placed instead. Any
number of abstract methods will be coalesced with a single concrete
method, but conflicts are always propagated unless the method is
overridden locally.