Capture only self-requests known at point of call
A receiverless request in a method or object constructor body should be
made on the containing object only if the method in question was
available at the level of hierarchy the code is at. Simply inserting the
final object into the scope stack leads to unanticipated downcalls,
which is incorrect behaviour. As part-objects are now verboten, the
quasi-static scoping must be simulated. A new SelfRedirectorObject type
will have an instance in the scope stack, which knows about the set of
methods found at the relevant level and captures any requests to send to
the self-object, without that object having to be in the scope stack
itself.