Add builtins extension function for all objects
Similarly to the functionality for strings and numbers, it will be
possible to extend the methods available on all objects (together) by
defining the ObjectExtension trait. The mechanism is different than for
the builtins, and uses a fallback functionality at the top of the tree,
meaning that these methods are available even on objects that existed
prior to the evaluation of the trait.