Allow class names to be operators
Previously, class names were enforced as starting with an identifier,
eliminating operators from contention. This requirement was too strict,
as operators are valid method names and should be available for classes,
so "class +(other) { ... }" will now be accepted. The same error will be
raised if neither an identifier nor operator is provided.