Bind "self" across methods and initialiser code
Previously, the binding of "self" for a method body was determined
dynamically when the method was called. A separate scope binding "self"
for the initialisation code of the object body was also defined, after
methods had already been constructed, and so methods were not inside
this scope. Other local bindings, such as imports, would be placed into
this initialiser scope, meaning methods did not have access to them.

Now, a single interior scope for a given object constructor is created
at construction time, and both "self" and any imports are bound into
that scope. Methods are defined within that interior scope, which is
placed inside the redirector scope used for implicit-receiver calls. All
code in an object body executes within the same scope, whether inside or
outside of a method.