Add bidirectional WebSocket RPC and execution
The --ws option to the console entry point launches a WebSocket server,
which can be connected to to execute code and receive output. The
websocket/dom module gives access to call functions and access objects
on the browser side, and converts Grace objects to the corresponding JS
types, JS objects to Grace types or to wrapped RPC objects, and blocks
to functions that make callbacks.