Add separate WS page and websocket/server module
The websocket/server module can be imported to convert a program running
at the console into one that can communicate over the web socket. The
wsclient/index.html page can be loaded to access the other end of the
socket independently of any IDE. In this way, regular programs can be
run from actual files.