Support [ ... ] parameters and remove indexing
This commit removes postcircumfix ("indexing") operator support, and
parses these cases instead as circumfix-bracket singleton arguments to
the method, in the same manner as strings or blocks.

The existing circumfix bracket operator support is used, and
circumfix[*x] is defined to construct an iterable in the prelude. Other
bracket operators are, for the moment, still supported by user
definition.

A new test case t331_lineup verifies the standard-compliant part of this
behaviour.