Remove bindings from matching infrastructure
All destructuring infrastructure is now removed, including all uses of
the "bindings" method in the system. This is a behaviour-altering
change.

Because there is nowhere values are threaded through, a variable is
always bound to the result of the last-matched pattern. In the case of
the & combinator, this is always the right-hand side, but for the |
combinator it is the result of the first pattern that matches. This
affects all uses of & and | with patterns that return results that are
not aliases of the target.

Other still-valid code should not change in behaviour.