Enable overloading by arity
Methods with differing numbers of parameters, but the same names, will
henceforth be treated as different methods. Arities are represented on
each part of the name as a comma-separated list of _ (underscore)
characters in parentheses after the part name, unless there are zero
arguments or parameters, in which case there is no indication. Parts
continue to be separated by spaces. Methods are identified by these
names.

The declaration:

    method foo(x,y) bar(z) { ... }

creates a method called:

    foo(_,_) bar(_)

Binary operators are also subject to this indication, for some reason.
Prefix operators are also, but the indication is always empty for these
methods. Binding methods continue to have an empty first part, but the
second part will be ":=(_)".

The bulk of updates in this commit are changes to the inclusion of
built-in methods, as all such methods taking parameters must now have
their parameters listed in their names.

Circumfix methods are not subject to arity overloading. Only one
circumfix method with any pair of brackets is available, and they
continue to have access to variadic parameters.