Add dom.yieldFor combining sleep + yield
The dom.sleep method blocks for some period of time, and dom.yield
flushes the event loop. The new yieldFor method combines the behaviour
of the two: it pauses the flow for some number of milliseconds before
continuing, but consumes any events that occur during that time and
processes them as yield would have. This can create a smoother
interactive experience.