Return to dynamically binding receiver in methods
When the receiver of requests was bound in scope at object construction
time, a consistent view was provided to all code in the object, but the
object extension traits were given the wrong idea of who they were
dynamically. Because the builtins are not user object, inheriting the
trait dynamically with each creation is not an option, so reverting this
aspect (while leaving the other scope changes that went with it in
place) is the simplest way to allow builtin extensions to work again.