Compute identifying name of aliases
New alias names were not put through the translation process to the
execution tree, and were given non-canonical names (making the methods
inaccessible through ordinary calls). This commit transforms the parse
nodes for new names in the same way as the nodes for old names,
retrieves the canonical name from the new node for binding, and then
discards the new node, which has no further role to play.

This corrects the behaviour of aliases with method parameters.