Permit dialects to record many errors
The Grace objects for AST nodes gain two new methods:
    report(code, message)
    report(code, message) with(vars)
These methods record an error with the given code and default message at
the location of the AST node. The "with(vars)" version requires a lineup
of lineups of strings, which are key-value paired up for string
replacements in the message ("${variable replacement}").