Process := in expression instead of statement
It was not previously possible to use a bind as an expression directly.
Code such as
    (x := 1)
would be a parse error, as would
    x(y := 2)
because := was only processes in non-trait statement context. With this
commit, binds are processed as part of ordinary expression parsing.

There is a new special case for locations that should not have this
behaviour: type and other annotations. "var x : A := 1" should not treat
"A:=1" as a request, and nor should "var x is readable := 1". In all
other places a general expression is expected (that is, where a
following operator is permitted) x := ... will be treated as accessing
the writer.