Fix pretty-print crash from API change
The MethodRequest API requires that request parts have their complete
argument list provided when created in order to generate the correct
arity-overloaded name for the method, but the pretty-printing code
modified a part after the fact, leading to a crash. This commit changes
the pretty-printing native code to use a different API that constructs
the request correctly.