Replace use of Enumerable.Repeat for JSIL
JSIL currently does not support Enumerable.Repeat, which is used in only
one place in the Kernan codebase. To maintain support with the current
master branch of JSIL, this commit introduces a manual loop building up
a string directly in place of that call.

A JSIL pull request is open and this change can be reverted if it is
merged.