Avoid cyclic import error when installing module
Previously, interactively installing a remote module would succeed, but
then report an error that the import chain was cyclic because it would
immediately recurse, having already marked the path as used. Running the
same code again would succeed as the module was correctly installed
before the error.

Now, the imported path is forgotten before the recursive step and the
program can continue. The first execution will work as well as
subsequent runs.