Correct WS parse-tree export of comments
The recursion within makeNode for comment chains had multiple errors:
first, the wrong overload of Visit would be called, and after that it
would attempt to add an element as its own descendant. This commit
corrects both errors and allows comments, including multi-comment
chains, to be transformed into JSON as other nodes are.