Correct indent inside block beginning with comment
Previously, this code:

  if (true) then { // always runs!
    1
    2
  }

would give an indentation error on "2" because the indentation level was
set to the offset of the comment (while the first body token got a free
pass). Now, the comment will be ignored for the purposes of computing
the indent level.

Further fixes for indentation handling at the beginning of blocks will
likely be required, particularly to deal with the speculative parsing of
parameters.